% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagedoc.R
\docType{package}
\name{oureuc}
\alias{oureuc}
\alias{oureuc-package}
\title{eucpackage: A package for computating the greatest common divisor of two numbers.}
\description{
The euclidian package provides one function called euclidian.
}
\section{euclidianpackage function}{
it takes two numbers,
  and returns thier greatest common factor
}

