% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikigraph is the data frame of three vectors, containing values in order to plot graph.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{V1,V2}{Representing the nodes of V1 and V2}
  \item{W}{Weight of the corresponding nodes for each observation}
  ...
}}
\usage{
wiki_graph
}
\description{
Wikigraph is the data frame of three vectors, containing values in order to plot graph.
}
\keyword{datasets}

